+++ psistats/src/main.rs
@@ -2,5 +2,5 @@ use psistats::service;
 #[macro_use] extern crate log;
 
 pub fn main() {
-    service::main::main();
+    service::main::cli_main();
 }
