+++ psistats/src/plugins/loader.rs
