+++ psistats/src/plugins/mod.rs
