+++ psistats/src/plugins/registry.rs
