+++ psistats/src/service/manager/publishers.rs
