+++ psistats_plugins/cpu/Cargo.toml
@@ -11,7 +11,7 @@ crate-type = ["cdylib"]
 name = "plugin_cpu"
 
 [dependencies]
-psistats = { path = "../../psistats_service" }
+psistats = { path = "../../psistats" }
 crossbeam-channel = "0.4.2"
 lazy_static = "1.4.0"
 sysinfo = { git = "https://github.com/GuillaumeGomez/sysinfo" }
