+++ psistats_plugins/logger/Cargo.toml
@@ -11,7 +11,7 @@ crate-type = ["cdylib"]
 name = "plugin_logger"
 
 [dependencies]
-psistats = { path = "../../psistats_service" }
+psistats = { path = "../../psistats" }
 log = "0.4.8"
 pretty_env_logger = "0.4.0"
