+++ wix/psistats-service-winsw.exe
